Deported request test function in helper
